ref(swap): move simialr logic into smaller components
